requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
openai
groq
pandas
python-dotenv
grpcio
grpcio-tools
streamlit